Fixes an issue with server parsing, connects to the servers
This fixes a bug where the parsed servers weren't stored locally. This
issue existed because I iterated by value instead of by reference,
causing the objects stored in the list to not be updated.

Additionally, this is the first update where the project is actually
useful. It now has a replica of socket.emit, as well as socket.on, and
it should be able to handle most parsing. It also connects to the point
where it receives the entities event, which is the next step to login.

There's also some new docs added in this version, but I hope to get
further into these when I'm closer to finishing this project.