Heavy refining of the core system
Adds UBSan suppression for uvw (false positives in resources.hpp).
Partial of #3 (implements auto-reconnect for wait_n_seconds). This
commit also significantly improves stability, by adding an intermediate
map. This map is modified directly by the socket, and socket
modification is only blocked when the map is copied and cleared. This
enables less blockage for the system, as well as potentially
obliterating data races for the entities map for the vast majority of
processes.

The libuv loop the timers run on, including the one that now processes
the entities map by adding from and clearing an intermediate entity map,
blocks execution of other loops and timers until it finishes, because
the libuv loop isn't async. Running at a speed of 60 FPS, it shouldn't
be too blocking for other processes either. This also aligns with the
standard game speed.

Finally, the interface has been improved. setMainLoop() is no more! That
system was rather horrible anyway. Use setInterval() instead. 1